+++ client/src/features/archive/index.tsx
@@ -1,12 +1,13 @@
 import { Route, Routes } from 'react-router-dom';
 import Overview from './overview';
+import ViewArchive from './view';
 
 export const Archive = () => {
 
   return (
     <Routes>
-      <Route path='*' element={<Overview/>} />
-      <Route path='/:archiveID' element={<></>} />
+      <Route path='*' element={<Overview />} />
+      <Route path='/:archiveID' element={<ViewArchive />} />
     </Routes>
   );
 };