+++ client/src/types/IDB_Archive.ts
@@ -1,10 +1,12 @@
+import IDB_File from "./iDB_file";
+
 export type IDB_Archive = {
   id: string;
   data: {
     title: string;
     report: string;
     description?: string;
-    files?: File[];
+    files?: IDB_File[];
   };
   meta: {
     authorID: number;